Переименовывает файлы.
