+++ project/src/store/offer-process/selectors.ts
@@ -1,7 +1,9 @@
-import { NameSpace, SortTypes } from '../../const';
-import { City } from '../../types/offers';
+import { NameSpace } from '../../const';
+import { Offer, Offers } from '../../types/offers';
 import { State } from '../../types/state';
 
-export const getCity = (state: State): City => state[NameSpace.Offer].city;
-export const getSortOptionOffers = (state: State): SortTypes => state[NameSpace.Offer].sortOptionOffers;
-export const getsetSelectedOfferId = (state: State): number | undefined => state[NameSpace.Offer].selectedOfferId;
+export const getOffers = (state: State): Offers => state[NameSpace.Data].offers;
+export const getDataLoadingStatus = (state: State): boolean => state[NameSpace.Data].isDataLoading;
+export const getCurrentOffer = (state: State): Offer | undefined => state[NameSpace.Data].currentOffer;
+export const getNearbyOffers = (state: State): Offers => state[NameSpace.Data].nearbyOffers;
+export const getErrorStatus = (state: State): boolean => state[NameSpace.Data].hasError;
